Improvements: Make repo optional, updata naming conventions

+++ scripts/handler.sh
@@ -1,33 +1,22 @@
 #!/bin/bash
 set -e
 
-function branchEnv()
-{
-	if [[ $1 == "epel8" ]]
-	then
-		RET="rocky+epel-8-x86_64"
-	else
-		[[ $1 =~ ^f[0-9]+$ ]] && RET="fedora-$(echo $1 | tr -d [a-z])-x86_64"
-	fi
-	echo $RET
-}
-
 function build()
 {
 	pkg=$1
-	branch=$2
-	repo=$(cat packages/$pkg/repo | tr -d " " | tr -d "\n")
-	mock_env=$(branchEnv $branch)
-	[[ ! -d packages/$pkg/$pkg ]] && git clone --recursive $repo packages/$pkg/$pkg &>>${LOGS_DIR}/${pkg}.log
-	git --git-dir=packages/$pkg/$pkg/.git --work-tree=packages/$pkg/$pkg reset --hard origin/$branch &>>${LOGS_DIR}/${pkg}.log
+	branch=$(echo $2 | cut -d ":" -f1)
+	mock_env=$(echo $2 | cut -d ":" -f2)
+	[[ -e packages/$pkg/repo ]] && repo=$(cat packages/$pkg/repo | tr -d " " | tr -d "\n")
+	[[ -e packages/$pkg/repo ]] && [[ ! -d packages/$pkg/$pkg ]] && git clone --recursive $repo packages/$pkg/$pkg &>>${LOGS_DIR}/${pkg}.log
+	[[ -e packages/$pkg/repo ]] && git --git-dir=packages/$pkg/$pkg/.git --work-tree=packages/$pkg/$pkg reset --hard origin/$branch &>>${LOGS_DIR}/${pkg}.log
 	release=$(grep 'Release:' packages/$pkg/$pkg/$pkg.spec | cut -d ":" -f2  | tr -d " " | cut -b -1)
 	version=$(grep 'Version:' packages/$pkg/$pkg/$pkg.spec | cut -d ":" -f2  | tr -d " ")
 	name="${pkg}-${version}-${release}.fc34.src.rpm"
 	echo "#########################################################################"
 	echo "### Building $pkg => $name"
 	echo "### Mockenv: $mock_env"
+	echo "### Branch: $branch"
 	echo "### Logs in $name.log"
-	echo "#########################################################################"
 	docker run -it -d --name ${pkg}-${branch}-build -v /home/circleci/project:/project --cap-add=SYS_ADMIN --security-opt apparmor:unconfined spideyz0r/mockzor &>>${LOGS_DIR}/${pkg}.log
 	echo "### Building SRPM for $pkg"
 	docker exec ${pkg}-${branch}-build /project/scripts/buildsrpm.sh $pkg &>>${LOGS_DIR}/${pkg}.log
@@ -44,9 +33,9 @@ packages=$(git diff main... --diff-filter=MACR --name-only | grep "packages/" | 
 
 for pkg in $packages
 do
-	for distro in $(cat packages/${pkg}/branches)
+	for env in $(cat packages/${pkg}/env)
 	do
-		build $pkg $distro
+		build $pkg $env
 	done
 done
 # wait
